Downgrade Play to 2.x to stick to Scala 2.12 across the repo (#43)
## Summary
Downgrade Play to 2.x. The version of Play we had required Scala 2.13.
This ends up being painful to work with when we start introducing
dependencies from the 'hub' play module to other modules like api /
online (which we need for the Dynamodb hookup). Sbt isn't very happy
when we're trying to hook up these builds across scala versions when hub
is defaulting to 2.13 and the rest of the project on 2.12. To not bang
at this further, I'm downgrading to a play version on scala 2.12 and we
can revisit when we're ready to bump to Scala 2.13.

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI
- [X] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced mock data generation for time series, allowing for a broader
range of generated values.
- Updated Java options in the Docker initialization script to prevent
execution errors.

- **Bug Fixes**
- Improved error handling in test cases by correctly accessing the
successful results from decoded responses.

- **Chores**
- Updated project dependencies and plugins to maintain compatibility and
improve build configuration, including the Play Framework plugin and
several SBT plugins.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->