Revert "Replace `npm` with `pnpm` (#136)" (#144)
## Summary

Reverting back to `npm` due to Snyk UI not supporting `pnpm.overrides`
in `package.json` (best source appears to be this
[comment](https://github.com/snyk/nodejs-lockfile-parser/issues/111#issuecomment-2384915606)).
Sounds like the CLI might support it at this time, which means the Snyk
[github action](https://github.com/snyk/actions/tree/master/node) might
work, but not worth exploring further at this time. Also not sure if
Vanta requires Snyk UI for compliance.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Co-authored-by: Sean Lynch <techniq35@gmail.com>