chore: Refactor out Format abstractions (#145)
## Summary
- Step 0 of https://app.asana.com/0/1208949807589885/1208951092959581/f 
- The end goal of this PR stack will be to support spark reading from
and writing to BigQuery. To do this we'll leverage the `Format`
abstraction to delegate out certain catalog operations such as listing
partitions. I will do a followup of this to introduce
`BigQueryFormatProvider` that will cover the bigquery cases.
- This PR is simply just factoring out `Format`, and also modifying its
class signature and cleaning up the reflection classloading.


## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a trait-based architecture for managing various table
formats in Apache Spark.
	- Added support for dynamic loading of format providers at runtime.

- **Bug Fixes**
- Enhanced partition retrieval logic for Iceberg and DeltaLake formats
to address known issues.

- **Refactor**
- Streamlined the `TableUtils` class and improved error handling for
unsupported operations.
- Removed redundant code and commented sections for better
maintainability.

- **Tests**
- Added a new test case for dynamic class loading of the table format
provider.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->