Add `npm run up-deps` and update deps to latest (Vite 5=>6, Svelte, etc) (#150)
## Summary

Due to some of our odd issues (out of memory, etc) I thought it would be
helpful to be on the latest Vite and Svelte versions.

I also added `npm run up-deps` similar to pnpm PR but leveraging
[npm-check-updates](https://github.com/raineorshine/npm-check-updates)
package to make bumping versions easier (and interactive).

This updates all dependencies except `bits-ui` (0.21.16 => 0.22.0) and
`svelte-radix` (1.1.1 => 2.0.1) which are used by shadcn-svelte. They
are probably safe to update (especially bits-ui) but would require more
visual validation.

I had to add a `vite` override as some internal deps of `vitest` were
still requesting `vite@5` (and oddly this isn't an issue when using
`pnpm` as I have a similar vite 6 setup on other repos.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
	- Added a new script for dependency management.

- **Updates**
- Multiple package versions updated to enhance performance and
compatibility, including key libraries such as Svelte, Playwright, and
TypeScript.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Sean Lynch <techniq35@gmail.com>