Add support to register UDFs in Flink (#142)
## Summary

Our Spark on Flink code doesn't include registering UDFs which is a gap
compared to our Spark structured streaming implementation. This PR adds
support for this. I've skipped registering of UDFs in derivations - can
add this either as part of this PR or in a follow up when the need
arises.

To confirm the jar registration works I added a directory in quickstart
with a couple of example UDFs and register that jar in my
SparkExprEvalFnTest to confirm things work as expected.

## Checklist
- [X] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced functionality for evaluating Spark SQL expressions with
additional setup configurations.
	- Introduced new user-defined functions (UDFs) for testing purposes.
	- Added support for Hive UDF registration in the `CatalystUtil` class.

- **Bug Fixes**
	- Improved error handling for setup statements in `CatalystUtil`.

- **Tests**
- Added new test methods to validate UDF functionality and integration
with the `CatalystUtil` framework.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->