feat: CatalogAwareDataPointer and refactoring existing DataPointer (#157)
## Summary

- Refactor DataPointer. Ideally, if we are parsing URI's, if we come
across a prefix we should preserve the original whole prefix. We lose
out on the benefits of the various `s3<c>` uri's but we can fix that in
a future iteration. This way, the Extensions code is simpler.
- Define some DataframeWriter and DataframeReader implicit classes to
support handling DataPointer. Ideally we want DataPointer to be a
lightweight object that we can take action on, similar to what a table
name or a uri is.
- Introduce CatalogAwareDataPointer. The way this works is that it
encapsulates `Format` which is a runtime injection used to figure out
underlying storage r/w layers. This is ultimately what DataPointer
represents, and instead of statically defining it we will make remote
calls to do so.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Release Notes

- **New Features**
- Introduced a more flexible `DataPointer` architecture with an abstract
base class and `URIDataPointer`.
  - Added support for dynamic format resolution in Spark data sources.
  - Enhanced `BQuery` and `GCS` classes with specific `name` methods.

- **Refactor**
  - Restructured `DataPointer` class to improve extensibility.
- Enhanced format handling with standardized `name` methods for
different data formats.
  - Updated `DataPointerOps` to streamline format and catalog handling.
- Modified `TableUtils` to utilize the new `DataPointer` instantiation
method.

- **Improvements**
  - Implemented more robust table and format parsing mechanisms.
  - Added utility methods for resolving table names and formats.
- Refined logging for `DataPointer` instantiation and state
representation.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->