Yank out Bigtable hbase dependency for one ApiFutureUtils class (#184)
## Summary
Been working through a number of issues trying to successfully run Flink
jobs on our DataProc cluster due to class / version conflicts. A number
of these seem to stem from this library that brings in a lot of
dependencies like hadoop ipc libs etc. We pulled in this lib for one
class (ApiFutureUtils) that allows us to convert from BigTable's
ApiFuture to Java's ListenableFutures. I've chosen to fork this class
and yank out the lib dependency.

## Checklist
- [ ] Added Unit Tests
- [X] Covered by existing CI - Our bigTableKVStoreTest runs tests for
the various async apis which leverage this class
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

## Release Notes

- **Dependency Management**
- Removed Google Cloud Bigtable HBase library dependency from project
configuration

- **New Features**
- Added utility method to convert Google Cloud API futures to Java
CompletableFutures

- **Refactor**
  - Removed specific Bigtable API utility import from implementation

<!-- end of auto-generated comment: release notes by coderabbit.ai -->