feat: additional date parsing and write confs (#181)
## Summary

- do some date formatting because the input partitions may be of `DATE`
type, so we need to normalize them to a string.
- inline some options for bigquery writes for now, will follow up with
more abstracted handling. See:
https://github.com/zipline-ai/chronon/pull/175
## Checklist
- [ ] Added Unit Tests
- [x] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Updated partition column and format handling in DataFrame processing.
- Modified visibility of `partitionFormat` to improve package-level
access.
  - Removed optional write format variable from TableUtils.

These changes enhance the functionality of data processing utilities by
refining how partition columns are managed and accessed within the Spark
extensions.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>