feat: read and write date types (#192)
## Summary

In BigQuery, time or date based partition columns cannot be written as
just STRING and must be DATE or TIMESTAMP. To support this, let's just
try to make this convention consistent across the board - this should be
doable for all other existing formats. It however may be a breaking
change so let's see if existing unit tests run successfully.



## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
	- Added configuration for temporary GCS bucket in GCP integration.
	- Enhanced BigQuery write format with additional configuration options.
	- Introduced new `options` method in Format trait.

- **Improvements**
- Updated data handling in Spark joins to improve partition column
formatting.
- Modified data pointer options retrieval to combine read and write
options.
- Refined BigQuery write method handling by removing the indirect
option.

- **Configuration Changes**
- Added temporary GCS bucket configuration:
`spark.chronon.table.gcs.temporary_gcs_bucket`.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->
---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208949807589898

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>