fix: back to thrift-typescript package (#211)
## Summary

While working with the types generated by the Thrift compiler, I noticed
it was generating files using the `require` keyword. This is only
compatible with a Node.js server and not usable in the browser. I
attempted several fixes but couldn't find a solution.

Ultimately, I switched to the `thrift-typescript` package and managed to
get it working. As of now, it provides everything we need for the
exploration code.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Dependencies**
	- Added `@creditkarma/thrift-server-core` (v1.0.4)
	- Added `@creditkarma/thrift-typescript` (v3.7.6)
	- Added `glob` (v9.0.0)

- **Chores**
	- Updated code generation script to improve Thrift file processing
	- Adjusted directory paths for Thrift file generation
	- Added command to clear output directory before generating new files
<!-- end of auto-generated comment: release notes by coderabbit.ai -->