fix: handle partition overwrite (#206)
## Summary

Based on 
[@David Han](https://zipline-2kh4520.slack.com/team/U0846REC8F7)’s
observation's around the spark-bigquery
[connector](https://zipline-2kh4520.slack.com/archives/C08710CDH8D/p1736700947319249?thread_ts=1736644291.357239&cid=C08710CDH8D),
there indeed is a lurking behavior.
when creating BQ tables (in the case they don't exist), user needs to
specify a partitioning. This is expected. We do that in the form of a
partitionColumn write option.
when the connector performs dynamic partition overwrites, you don't need
to specify the partitioning at all. It will do the right thing because
the destination table was already created with a partition spec.
^ ideally, the above could be idempotent even if the user passes the
partition column to the write, but unfortunately it's a strict
requirement that you don't define the partition column when doing
dynamic partition overwrites.
Fix is to specify the partition column only when the table DNE and needs
to be created, and leave it out in all other cases.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
	- Introduced separate read and write options for data pointers
	- Added support for Google Cloud SDK tool
	- Added new plugins for `gcloud` and `thrift`

- **Bug Fixes**
	- Enhanced error handling in BigQuery format provider

- **Refactor**
- Standardized `DataPointer` instantiation method from `apply` to `from`
	- Improved options handling in data operations

- **Chores**
	- Updated plugin and tool versions for development environment

<!-- end of auto-generated comment: release notes by coderabbit.ai -->
---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209143482009688

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>