Bazel migration for spark module (#196)
## Summary
Migrated spark module to bazel.

Known issue with tests : JUnit and Scala tests using FunSuite are not
getting identified in bazel scala_test_suite. Thanks to @nikhil-zlai for
working on the script to modify our tests to use Scala FlatSpec api
which seems to be working with bazel. Will test again after his changes
and create a separate PR if additional changes are needed. Also, it's
good to be consistent with the unit tests api going forward unless we
really need others.

## Checklist
- [ ] Added Unit Tests
- [x] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Release Notes

- **Dependencies**
- Updated logging framework dependency to Log4j 2.x across various
libraries.
	- Updated ScalaTest dependencies in API testing.
	- Migrated library dependencies from Maven to Scala artifacts.
	- Updated Jackson, JSON4s, and other library versions.
	- Added new dependencies for Guava, Netty, Delta Spark, and Kafka.

- **Build Configuration**
	- Introduced new Scala library and test configurations.
	- Added JVM binary assembly configuration.
	- Refined Bazel build rule imports and loading.

- **Dependency Management**
	- Removed several unused dependencies.
	- Simplified dependency references.
	- Updated runtime dependency paths.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->