feat: add a workflow to poll other workflows for branch protection (#289)
## Summary

- branch protection requires explicitly adding github actions that need
to succeed before a user can merge a PR. Some github workflows do not
qualify to run for a PR based on path filtering etc. Skipped workflows
will still be required to "succeed" for a branch protection rule. This
PR adds a blanked workflow that will poll for workflows explicitly
triggered to succeed instead.

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
- Added a new GitHub Actions workflow to enforce status checks on pull
requests.
- Reformatted test file for improved code readability without changing
functionality.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>