Flink/Fetcher tile key thrift interface (#300)
## Summary
Define the thrift interface for the TileKey struct as we're building out
the tiling + tile layering implementation on GCP.

Idea is that Flink code will create a tile key struct with the
appropriate fields, serialize it and send it as part of a PutRequest
object in a multiput call. BigTableKVStoreImpl (and later DDBImpl etc)
will deser this TileKey struct, construct the appropriate RowKey /
timestamp structure and write out to the store.

On the Fetcher side for each tile size we support (just one until we
support tile layering) we will create a TileKey and pass it as part of a
GetRequest object (which also includes the start / end range). KV store
will deser this and run the lookup + range query.

I've currently not added utils to help with murmur hashing pieces of the
TileKey that are needed for BigTable's RowKey. We can add those if we
decide to pursue the hashing to distribute data option.

## Checklist
- [X] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced `TilingUtils` with serialization and deserialization
capabilities for `TileKey` objects.
	- Added `TileKey` struct for managing tile-based data partitioning.

- **Tests**
- Added comprehensive test suite for `TilingUtils` serialization
functionality.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->