feat: parquet options intermed format (#361)
## Summary

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced data processing by introducing new configuration options for
writing data, including support for Parquet as the intermediate format
and enabling list inference during write operations.
- Expanded selection of fields in purchase events with the addition of
`bucket_rand`.
- Introduced a new aggregation to calculate the last 15 purchase prices,
utilizing the newly added `bucket_rand` field.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>