feat: make the thrift get cross platform compat (#386)
## Summary

- Make the thrift gen python executable, use `py_binary` to support
python generally

## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update
<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
- Enhanced the build process for a key automation tool by streamlining
its execution and command handling, leading to improved overall build
reliability and performance.
- Transitioned the export mechanism of a Python script to a defined
executable binary target for better integration within the build system.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Co-authored-by: Thomas Chow <thomaschow369@gmail.com>