Revert "feat: bump spark to 3.5.4 (#392)" (#395)
## Summary

- Even though I'm eager to get ahead here, let's not go too crazy and
accidentally shoot ourselves in the foot. Let's stay pinned to what our
clusters have (3.5.1) until those upgrade.



## Checklist
- [ ] Added Unit Tests
- [ ] Covered by existing CI
- [ ] Integration tested
- [ ] Documentation update


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated core Spark libraries—impacting SQL, Hive, Streaming, and Avro
features—to version 3.5.1 to ensure enhanced stability and improved
integration across Spark-powered functionalities.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->